requirements.txt
requests
python-vlc
pytube